Replace QuestionTable with new data-table
